+++ src/ConstantQ.cpp
@@ -78,11 +78,11 @@ ConstantQ::getMinFrequency() const
 }
 
 double
-ConstantQ::getBinFrequency(int bin) const
+ConstantQ::getBinFrequency(double bin) const
 {
     // our bins are returned in high->low order
     bin = (getBinsPerOctave() * getOctaves()) - bin - 1;
-    return getMinFrequency() * pow(2, (double(bin) / getBinsPerOctave()));
+    return getMinFrequency() * pow(2, (bin / getBinsPerOctave()));
 }
 
 void
